chore: dedupe backup retention and spy 429 throttle notify
Share getBackupRetention clamp/settings read for manual and auto prune paths, and assert notifyRateLimited is called from existing provider 429 tests.
